# Implement Cairo0 Class Hash