--- a/clients/feeder/feeder.go
+++ b/clients/feeder/feeder.go
@@ -11,6 +11,7 @@ import (
 	"os"
 	"path/filepath"
 	"strings"
+	"testing"
 	"time"
 
 	"github.com/NethermindEth/juno/core/felt"
@@ -91,7 +92,7 @@ func NopBackoff(d time.Duration) time.Duration {
 }
 
 // NewTestClient returns a client and a function to close a test server.
-func NewTestClient(t utils.BenchmarkTesting, network *utils.Network) *Client {
+func NewTestClient(t testing.TB, network *utils.Network) *Client {
 	srv := newTestServer(t, network)
 	t.Cleanup(srv.Close)
 	ua := "Juno/v0.0.1-test Starknet Implementation"
@@ -116,7 +117,7 @@ func NewTestClient(t utils.BenchmarkTesting, network *utils.Network) *Client {
 	return c
 }
 
-func newTestServer(t utils.BenchmarkTesting, network *utils.Network) *httptest.Server {
+func newTestServer(t testing.TB, network *utils.Network) *httptest.Server {
 	return httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		queryMap, err := url.ParseQuery(r.URL.RawQuery)
 		if err != nil {