--- a/core/rust/.gitignore
+++ b/core/rust/.gitignore
@@ -1,9 +1,0 @@
-/target
-
-
-# Added by cargo
-#
-# already existing elements were commented out
-
-#/target
-/Cargo.lock